비밀번호 확인 후 수정 및 삭제

+++ src/main/java/com/sparta/postboard/service/PostBoardService.java
@@ -33,16 +33,36 @@ public class PostBoardService {
     @Transactional
     public Long updatePostBoard(Long id, PostBoardRequestDto requestDto) {
         PostBoard postBoard = findPostBoard(id);
+        String pwd =  postBoard.getPassword();
+        String checkPwd =  requestDto.getPassword();
+        System.out.println("pwd => " + pwd);
+        System.out.println("checkPwd => " + checkPwd);
 
-        postBoard.update(requestDto);
+        if (pwd.equals(checkPwd)) {
+            postBoard.update(requestDto);
+            System.out.println("비밀번호 일치로 수정 성공");
+        } else {
+            System.out.println("비밀번호 불일치로 수정 실패");
+        }
+
 
         return id;
     }
 
-    public Long deletePostBoard(Long id) {
+    public Long deletePostBoard(Long id, PostBoardRequestDto requestDto) {
         PostBoard postBoard = findPostBoard(id);
+        String pwd =  postBoard.getPassword();
+        String checkPwd =  requestDto.getPassword();
+        System.out.println("pwd => " + pwd);
+        System.out.println("checkPwd => " + checkPwd);
 
-        postBoardRepository.delete(postBoard);
+        if (pwd.equals(checkPwd)) {
+            postBoardRepository.delete(postBoard);
+            System.out.println("비밀번호 일치로 삭제 성공");
+        } else {
+            System.out.println("비밀번호 불일치로 삭제 실패");
+        }
+
 
         return id;
     }
